requirements.txt
torch==1.7.1
torchvision==0.8.2
pillow==8.0.1
black==19.10b0
hydra-core==1.0
